Revert "created bogo build scripts and moved shim to examples"
