+++ openssl-test/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "openssl-test"
+version = "0.1.0"
+authors = ["SimranGujral <gujralsimran19@gmail.com>"]
+
+[dependencies]
+webpki = "0.12"
+env_logger = "0.4.2"
+openssl = "*"
+pem = "0.3"
