--- openssl-test/runme
@@ -1,19 +1,0 @@
-#!/bin/sh
-
-# This script fetches, builds, and runs the BoringSSL
-# TLS test tool ('BoGo') against rustls.
-
-set -xe
-
-if [ ! -e ../target/debug/examples/bogo_shim ] ; then
-  cargo test --no-run --features dangerous_configuration
-fi
-
-if [ ! -e bogo/ ] ; then
-  ./fetch-and-build
-fi
-
-( cd bogo && ./bogo.test -shim-path ../../target/debug/examples/bogo_shim \
-     -shim-config ../config.json \
-     -test.parallel 1 -num-workers 1 \
--pipe -allow-unimplemented )
