+++ openssl/Cargo.toml
@@ -28,8 +28,3 @@ tempdir = "0.3"
 winapi = "0.2"
 ws2_32-sys = "0.2"
 hex = "0.2"
-env_logger = "0.4"
-
-[[example]]
-name = "bogo_shim"
-path = "examples/bogo_shim.rs"
